Rename style file.
